% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifid.R
\name{bifid_matrix}
\alias{bifid_matrix}
\title{create a matrix for bifid handcipher}
\usage{
bifid_matrix(key)
}
\arguments{
\item{key}{a sentance or word that you use as a key}
}
\description{
create a matrix for bifid handcipher
}
\examples{
bifid_matrix("averysecretkey")
}
