% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifid.R
\name{encrypt_with_bifid}
\alias{encrypt_with_bifid}
\title{with a matrix return the encrypted word.}
\usage{
encrypt_with_bifid(bifidmatrix, phrase)
}
\description{
with a matrix return the encrypted word.
}
\examples{
encrypt_with_bifid(bifid_matrix,"puppersaresmalldoggos")
}
